+++ metadata-ingestion/src/datahub/ingestion/source/looker/lookml_source.py
@@ -90,7 +90,7 @@ from datahub.metadata.schema_classes import (
     SubTypesClass,
 )
 from datahub.utilities.lossy_collections import LossyList
-from datahub.utilities.source_helpers import (
+from datahub.api.source_helpers import (
     auto_stale_entity_removal,
     auto_status_aspect,
 )
