+++ metadata-ingestion/src/datahub/ingestion/source/pulsar.py
@@ -49,7 +49,7 @@ from datahub.metadata.schema_classes import (
     DatasetPropertiesClass,
     SubTypesClass,
 )
-from datahub.utilities.source_helpers import (
+from datahub.api.source_helpers import (
     auto_stale_entity_removal,
     auto_status_aspect,
 )
