+++ metadata-ingestion/src/datahub/ingestion/source/unity/source.py
@@ -84,7 +84,7 @@ from datahub.metadata.schema_classes import (
 )
 from datahub.utilities.hive_schema_to_avro import get_schema_fields_for_hive_column
 from datahub.utilities.registries.domain_registry import DomainRegistry
-from datahub.utilities.source_helpers import (
+from datahub.api.source_helpers import (
     auto_stale_entity_removal,
     auto_status_aspect,
     auto_workunit_reporter,
