+++ metadata-ingestion/tests/unit/test_source_helpers.py
@@ -3,7 +3,7 @@ from typing import List, Union
 import datahub.metadata.schema_classes as models
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
 from datahub.ingestion.api.workunit import MetadataWorkUnit
-from datahub.utilities.source_helpers import auto_status_aspect, auto_workunit
+from datahub.api.source_helpers import auto_status_aspect, auto_workunit
 
 _base_metadata: List[
     Union[MetadataChangeProposalWrapper, models.MetadataChangeEventClass]
